+++ Bakeries.DataAccess/Entities/PurchaseModel.cs
@@ -24,7 +24,7 @@ namespace Bakeries.DataAccess.Entities
     
 
         [Required]
-        public float Quantity { get; set; } // الكمية
+        public decimal Quantity { get; set; } // الكمية
 
         [Required]
         [StringLength(50)] // تحديد الحد الأقصى لعدد الحروف
